chore: rename tagControllers to eventTagControllers
